This is mostly working except for one weird UI bug:
When I go to the QuestionDetailPage.xaml, then down-vote a question, then pop the page, the brightened downvote button on the QuestionDisplayCard appears about a centimetre to the right of where it should.

I suspect that this is because the greyed-out one becomes invisible, but the card's layout doesn't get redrawn as a result of this invisibility.

Possibly the correct solution is to use a single button with a changing VisualState like this ToggleButton:
https://learn.microsoft.com/en-us/xamarin/xamarin-forms/user-interface/button